requirements.txt
Appium-Python-Client == 0.24
selenium==3.141.0
setuptools
pytest
allure-pytest==2.9.45